+++ packages/php-wasm/src/web/progress-monitoring/progress-observer.ts
