+++ packages/wordpress-playground/src/client-test.ts
@@ -1,10 +1,15 @@
 import * as Comlink from 'comlink';
 
+import { setupTransferHandlers } from '@wordpress/php-wasm'
+setupTransferHandlers();
+
 const iframe = document.getElementById('wp') as HTMLIFrameElement;
 iframe.src = "/wordpress.html";
 const playground = Comlink.wrap(Comlink.windowEndpoint(iframe.contentWindow!)) as any;
 await new Promise((resolve) => iframe.onload = resolve);
-
+await playground.onDownloadProgress(
+    Comlink.proxy(x => { console.log('download progress', {x}) })
+)
 console.log("Calling is ready")
 await playground.isReady();
 console.log("Called is ready")
