+++ packages/wordpress-playground/src/playground.html
@@ -10,8 +10,8 @@
         height: 100%;
       }
       iframe {
-        width: 100%;
-        height: 100%;
+        width: 100vw;
+        height: 100vh;
         border: none;
       }
     </style>
